+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 * When saving a Site you'll now see a success message
+* Disk usage of the root mountpoint is now shown in Stats Bar
 
 ### Changed
 * Main layout and stats bar completely redesigned